+++ strategies/__init__.py
@@ -10,6 +10,7 @@ from strategies.atm_straddle import atm_straddle
 from strategies.atm_straddle_index_move import atm_straddle_index_move
 from strategies.daily_put_sell import daily_put_sell
 from strategies.smoke_test_strangle import smoke_test_strangle
+from strategies.prod_test_put import prod_test_put
 
 __all__ = [
     "blueprint_strangle",
@@ -17,4 +18,5 @@ __all__ = [
     "atm_straddle_index_move",
     "daily_put_sell",
     "smoke_test_strangle",
+    "prod_test_put",
 ]